+++ backend/config.go
@@ -77,6 +77,8 @@ type Config struct {
 	NoEmbedEtcd                  bool
 	EtcdHeartbeatInterval        uint
 	EtcdElectionTimeout          uint
+	EtcdDiscovery                string
+	EtcdDiscoverySrv             string
 
 	// Etcd TLS configuration
 	EtcdClientTLSInfo     etcd.TLSInfo